Fix S.13 IPC hardening QA round 2 findings
