+++ crates/atm-daemon/src/composition.rs
@@ -3,6 +3,8 @@ use crate::boundary_adapters::{
     DaemonReconcileCoordinator, FileWatchEventSource,
 };
 use crate::daemon_runtime_observability::DaemonRuntimeObservability;
+use crate::host_ownership::HostOwnershipAdapter;
+use crate::local_ipc_transport::{RuntimeServeHooks, SocketEndpointGuard};
 use crate::runtime_health::DaemonRequestDispatcher;
 use crate::runtime_health::{DaemonStatusSource, RuntimeStatusCache};
 use crate::{
@@ -112,6 +114,8 @@ impl RuntimeLifecycle {
 #[derive(Debug)]
 pub(crate) struct RuntimeComposition {
     lifecycle: Arc<RuntimeLifecycle>,
+    _host_ownership: HostOwnershipAdapter,
+    endpoint_guard: Mutex<Option<SocketEndpointGuard>>,
     server_transport: LocalIpcServerTransportAdapter,
     request_dispatcher: Arc<DaemonRequestDispatcher>,
     _notification_sink: DaemonNotificationSink,
@@ -158,6 +162,8 @@ impl RuntimeComposition {
         };
         Ok(Self {
             lifecycle: Arc::new(RuntimeLifecycle::new()),
+            _host_ownership: HostOwnershipAdapter::new(),
+            endpoint_guard: Mutex::new(None),
             server_transport: LocalIpcServerTransportAdapter::new(),
             request_dispatcher: Arc::new(DaemonRequestDispatcher::new(
                 home_dir,
@@ -181,6 +187,30 @@ impl RuntimeComposition {
 
     fn request_dispatcher(&self) -> Arc<dyn RequestDispatcher + Send + Sync> {
         self.request_dispatcher.clone()
+    }
+
+    fn replace_endpoint_guard(&self, guard: Option<SocketEndpointGuard>) -> Result<(), AtmError> {
+        let mut slot = self.endpoint_guard.lock().map_err(|_| {
+            AtmError::daemon_unavailable("runtime endpoint guard slot lock poisoned").with_recovery(
+                "Restart the daemon; same-host endpoint cleanup ownership can no longer be tracked safely.",
+            )
+        })?;
+        *slot = guard;
+        Ok(())
+    }
+
+    fn take_endpoint_guard(&self) -> Result<SocketEndpointGuard, AtmError> {
+        let mut slot = self.endpoint_guard.lock().map_err(|_| {
+            AtmError::daemon_unavailable("runtime endpoint guard slot lock poisoned").with_recovery(
+                "Restart the daemon; same-host endpoint cleanup ownership can no longer be tracked safely.",
+            )
+        })?;
+        slot.take().ok_or_else(|| {
+            AtmError::daemon_unavailable("runtime endpoint guard was missing during daemon serve startup")
+                .with_recovery(
+                    "Restart the daemon; same-host endpoint cleanup ownership was lost before the listener entered serving state.",
+                )
+        })
     }
 
     fn begin_shutdown(&self) -> Result<(), AtmError> {
@@ -242,7 +272,7 @@ impl RuntimeComposition {
             self.lifecycle.force_stopped()?;
             return Err(error);
         }
-        let runtime = match self.server_transport.prepare_runtime() {
+        let mut runtime = match self.server_transport.prepare_runtime() {
             Ok(runtime) => runtime,
             Err(error) => {
                 if let Err(shutdown_error) = self.shutdown_background_lanes() {
@@ -260,6 +290,7 @@ impl RuntimeComposition {
                 return Err(error);
             }
         };
+        self.replace_endpoint_guard(Some(runtime.take_endpoint_guard()?))?;
         self.lifecycle.transition(RuntimeLifecycleState::Running)?;
         self.request_dispatcher.emit_runtime_event(
             "startup_completed",
@@ -267,13 +298,17 @@ impl RuntimeComposition {
             "daemon startup completed",
         );
         let request_dispatcher = Arc::clone(&self.request_dispatcher);
+        let endpoint_guard = self.take_endpoint_guard()?;
         let result = runtime.serve_with_runtime_hooks(
             self.request_dispatcher(),
-            super::GRACEFUL_DRAIN_DEADLINE,
-            super::FORCE_CANCEL_DEADLINE,
-            || self.begin_shutdown(),
-            move || request_dispatcher.reload_runtime_view(),
-            || self.finalize_shutdown(),
+            RuntimeServeHooks {
+                endpoint_guard,
+                graceful_drain_deadline: super::GRACEFUL_DRAIN_DEADLINE,
+                force_cancel_deadline: super::FORCE_CANCEL_DEADLINE,
+                begin_shutdown: || self.begin_shutdown(),
+                reload_runtime_view: move || request_dispatcher.reload_runtime_view(),
+                finalize_shutdown: || self.finalize_shutdown(),
+            },
         );
         self.finish_runtime(result)
     }
@@ -322,7 +357,7 @@ impl RuntimeComposition {
             self.lifecycle.force_stopped()?;
             return Err(error);
         }
-        let runtime = match self
+        let mut runtime = match self
             .server_transport
             .prepare_runtime_at_socket_path(socket_path)
         {
@@ -343,6 +378,7 @@ impl RuntimeComposition {
                 return Err(error);
             }
         };
+        self.replace_endpoint_guard(Some(runtime.take_endpoint_guard()?))?;
         self.lifecycle.transition(RuntimeLifecycleState::Running)?;
         self.request_dispatcher.emit_runtime_event(
             "startup_completed",
@@ -350,13 +386,17 @@ impl RuntimeComposition {
             "daemon startup completed",
         );
         let request_dispatcher = Arc::clone(&self.request_dispatcher);
+        let endpoint_guard = self.take_endpoint_guard()?;
         let result = runtime.serve_with_runtime_hooks(
             self.request_dispatcher(),
-            super::GRACEFUL_DRAIN_DEADLINE,
-            super::FORCE_CANCEL_DEADLINE,
-            || self.begin_shutdown(),
-            move || request_dispatcher.reload_runtime_view(),
-            || self.finalize_shutdown(),
+            RuntimeServeHooks {
+                endpoint_guard,
+                graceful_drain_deadline: super::GRACEFUL_DRAIN_DEADLINE,
+                force_cancel_deadline: super::FORCE_CANCEL_DEADLINE,
+                begin_shutdown: || self.begin_shutdown(),
+                reload_runtime_view: move || request_dispatcher.reload_runtime_view(),
+                finalize_shutdown: || self.finalize_shutdown(),
+            },
         );
         self.finish_runtime(result)
     }
